Revert "Separated to represent different Lambda functions, tweaked basic operations"
This reverts commit ffba4a77053d372b0401f44e6de00f09971d7b0a.
