+++ .gitignore
@@ -1,3 +1,2 @@
-centers/node_modules/
-donations/node_modules/
-users/node_modules/
+node_modules
+node_modules
