--- centers/models/Center.js
@@ -1,66 +1,0 @@
-module.exports = (sequelize, DataTypes) => {
-    const Center = sequelize.define("Center", {
-            id: {
-                type: DataTypes.STRING,
-                allowNull: false,
-                primaryKey: true
-            },
-            name: {
-                type: DataTypes.STRING,
-                allowNull: false,
-                validate: {
-                    notEmpty: true
-                }
-            },
-            description: {
-                type: DataTypes.STRING,
-                allowNull: false,
-                validate: {
-                    notEmpty: true
-                }
-            },
-            address: {
-                type: DataTypes.STRING,
-                allowNull: false,
-                validate: {
-                    notEmpty: true
-                }
-            },
-            latitude: {
-                type: DataTypes.FLOAT,
-                allowNull: false
-            },
-            longitude: {
-                type: DataTypes.FLOAT,
-                allowNull: false
-            },
-            email: {
-                type: DataTypes.STRING,
-                allowNull: false,
-                validate: {
-                    notEmpty: true
-                }
-            },
-            password: {
-                type: DataTypes.STRING,
-                allowNull: false,
-                validate: {
-                    notEmpty: true
-                }
-            },
-            profileImg: {
-                type: DataTypes.STRING
-            },
-            isEmailVerified: {
-                type: DataTypes.BOOLEAN,
-                defaultValue: false
-            },
-            isDocsVerified: {
-                type: DataTypes.BOOLEAN,
-                defaultValue: false
-            }
-        }
-    );
-
-    return Center;
-}
