--- centers/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "care-club-backend-centers",
-  "version": "1.0.0",
-  "description": "Backend for CareClub",
-  "main": "index.js",
-  "scripts": {
-    "test": "npm start",
-    "start": "nodemon index.js"
-  },
-  "author": "AWSAPU Team",
-  "license": "ISC",
-  "dependencies": {
-    "express": "^4.17.1",
-    "mysql2": "^2.2.5",
-    "sequelize": "^6.6.5",
-    "sequelize-cli": "^6.2.0",
-    "uniqid": "^5.3.0"
-  }
-}
