--- donations/models/Donation.js
@@ -1,44 +1,0 @@
-module.exports = (sequelize, DataTypes) => {
-    const Donation = sequelize.define("Donation", {
-        id: {
-            type: DataTypes.STRING,
-            allowNull: false,
-            primaryKey: true
-        },
-        name: {
-            type: DataTypes.STRING,
-            allowNull: false,
-            validate: {
-                notEmpty: true
-            }
-        },
-        description: {
-            type: DataTypes.STRING,
-        },
-        isPerishable: {
-            type: DataTypes.BOOLEAN,
-            allowNull: false,
-            defaultValue: false
-        },
-        quantity: {
-            type: DataTypes.INTEGER,
-            allowNull: false,
-        },
-        status: {
-            type: DataTypes.STRING,
-            allowNull: false,
-            defaultValue: "Pending"
-        }
-    })
-
-    Donation.associate = models => {
-        Donation.belongsTo(models.User, {
-            foreignKey: "userId",
-        })
-        Donation.belongsTo(models.Center, {
-            foreignKey: "centerId",
-        })
-    }
-
-    return Donation;
-}
