--- donations/models/User.js
@@ -1,65 +1,0 @@
-const { v1: uuidv1 } = require("uuid");
-
-module.exports = (sequelize, DataTypes) => {
-    const User = sequelize.define("User", {
-        id: {
-            type: DataTypes.STRING,
-            allowNull: false,
-            primaryKey: true,
-        },
-        firstName: {
-            type: DataTypes.STRING,
-            allowNull: false,
-            validate: {
-                notEmpty: true
-            }
-        },
-        lastName: {
-            type: DataTypes.STRING,
-            allowNull: false,
-            validate: {
-                notEmpty: true
-            }
-        },
-        email: {
-            type: DataTypes.STRING,
-            allowNull: false,
-            validate: {
-                notEmpty: true
-            }
-        },
-        password: {
-            type: DataTypes.STRING,
-            allowNull: false,
-            validate: {
-                notEmpty: true
-            }
-        },
-        phone: {
-            type: DataTypes.STRING
-        },
-        address: {
-            type: DataTypes.STRING
-        },
-        profileImg: {
-            type: DataTypes.STRING
-        },
-        isVerified: {
-            type: DataTypes.BOOLEAN,
-            defaultValue: false
-        },
-        points: {
-            type: DataTypes.INTEGER,
-            defaultValue: 0
-        }
-    })
-
-    User.associate = models => {
-        User.hasMany(models.Donation, {
-            onDelete: "cascade",
-            foreignKey: "userId"
-        })
-    }
-
-    return User;
-}
