--- users/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "care-club-backend-users",
-  "version": "1.0.0",
-  "description": "Backend for CareClub",
-  "main": "index.js",
-  "scripts": {
-    "test": "npm start",
-    "start": "nodemon index.js"
-  },
-  "author": "AWSAPU Team",
-  "license": "ISC",
-  "dependencies": {
-    "express": "^4.17.1",
-    "mysql2": "^2.2.5",
-    "sequelize": "^6.6.5",
-    "uuid": "^8.3.2"
-  }
-}
